requirements.txt
aiogram==2.25.1
motor==3.1.1
python-dotenv==1.0.0